refactor: api cleanup
