--- src/models/index.ts
@@ -1,4 +1,0 @@
-export { default as Account } from './Account'
-export { default as Progress } from './Progress'
-export { default as Data } from './Data'
-export { default as Feature } from './Feature'
